+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Allow `Entity` instances to be referenced weakly.
+
 ### Fixed
 - Added missing typing marker.
 - Corrected the type-hinting of `Entity.component.get` and `Entity.component.setdefault`.