+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Deprecated
+- `World.global_` has been deprecated since `world[None]` is simpler and less redundant.
 
 ## [3.0.0] - 2023-05-29
 ### Added